retarget method of CounterDef and TriggerDef into sub-module Counter and Trigger
